refactored marquee into typescript
